Fix ffmpeg GIF conversion command formatting
The ffmpeg command to convert GIFs to MP4s was not in back-ticks so it
was being rendered as markdown. This caused the command to render
incorrectly since the command contains asterisks which are significant
markdown syntax. This incorrect rendering means that you can no longer
copy & paste the example.